[Serializer] Moves dummy to fixtures folder
